agent: refactor config file parsing
*** BREAKING CHANGE ***

This change decouples the config file reading from the parsing to
make it easier to test and to support reading configuration from
other sources like environment variables. In the current implementation,
CONSUL_CONFIG is supported as an env var which can contain
configuration.

A side effect of the refactoring is that config files without a .json
suffix are no longer supported. This is in line with the currently
documented behavior but we need to decide whether we want to do this.