agent: decouple agent from api
*** FOR REVIEW ONLY: DOES NOT COMPILE ***

Making the agent depend on the api package for a couple of constants
wasn't such a good idea since it seems to prevent dropping the
dependency on the external binary for running the tests.

This patch reverses it but isn't complete yet. The TXN code requires
some more work and it was already necessary to copy a structure from the
api package.